@@ -2,23 +2,22 @@
 
 namespace esnerda\Json2CsvProcessor;
 
-use Keboola\Component\UserException;
+use Keboola\CsvTable\Table;
+use Psr\Log\LoggerInterface;
 use Symfony\Component\Finder\Finder;
 
 class Processor
 {
     const FILE_NAME_COL_NAME = 'keboola_file_name_col';
-    /** @var  string */
-    private $jsonParser;
+
+    private JsonToCSvParser $jsonParser;
     private $root_el;
 
-    /** @var  bool */
-    private $incremental;
+    private bool $incremental;
     private $add_row_nr;
     private $addFileName;
 
-    /** @var LoggerInterface */
-    private $logger;
+    private LoggerInterface $logger;
 
     public function __construct($jsonParser, bool $add_row_nr, bool $incremental, string $root_el, bool $addFileName, $logger)
     {
@@ -30,21 +29,20 @@ class Processor
         $this->logger = $logger;
     }
 
-    public function convert(string $datadir, string $type): self
+    public function convert(string $datadir, string $type): void
     {
-        return $this->processFiles(
-                        sprintf("%s/in/" . $type . '/', $datadir),
+        $this->processFiles(
+            sprintf("%s/in/" . $type . '/', $datadir),
             sprintf("%s/out/tables/", $datadir)
         );
     }
 
-    private function processFiles(string $inputDir, string $outputDir): self
+    private function processFiles(string $inputDir, string $outputDir): void
     {
         $finderFiles = new Finder();
         
         $finderFiles->files()->in($inputDir)->notName('*.manifest');
         $finderFiles->sortByName();
-        $manifests = $this->getManifests($inputDir);
 
         foreach ($finderFiles as $file) {
             $this->logger->info("Parsing file " . $file->getFileName());
@@ -58,13 +56,16 @@ class Processor
             if ($this->addFileName) {
                 $json_result_root = $this->addFileName(json_encode($json_result_root), $file->getFileName(), self::FILE_NAME_COL_NAME);
             }
-            $this->jsonParser->parse($json_result_root, $this->add_row_nr);
+
+            $this->jsonParser->parse($json_result_root);
         }
 
         $this->logger->info("Writting results..");
-        $csv_files = $this->jsonParser->getCsvFiles();
-        $this->storeResults($outputDir, $csv_files, $this->incremental);
-        return $this;
+        $this->storeResults(
+            $outputDir,
+            $this->jsonParser->getCsvFiles(),
+            $this->incremental
+        );
     }
 
     private function addFileName($json, $fileName, $colName)
@@ -96,41 +97,19 @@ class Processor
         }
     }
 
-    private function getManifests($inputDir)
-    {
-        $finderManifests = new Finder();
-        $manifests = [];
-        $finderManifests->files()->in($inputDir)->name('*.manifest');
-        foreach ($finderManifests->name('*.manifest') as $manifest) {
-            $manFile = file_get_contents($manifest->getRealPath());
-            $destination = explode('.', json_decode($manFile)->destination);
-            if (count($destination) >= 2) {
-                $manifests[$manifest->getFilename()] = $destination[0] . '.' . $destination[1];
-            }
-        }
-        return $manifests;
-    }
-
     /**
      * @param Table[] $csvFiles
-     * @param string $bucketName
-     * @param bool $sapiPrefix whether to prefix the output bucket with "in.c-"
-     * @param bool $incremental Set the incremental flag in manifest
-     * TODO: revisit this
      */
-    private function storeResults(string $outdir, array $csvFiles, $incremental = false, $bucketName = null, $sapiPrefix = true)
+    private function storeResults(string $outdir, array $csvFiles, bool $incremental = false): void
     {
         foreach ($csvFiles as $key => $file) {
             $path = $outdir;
 
             if ($file == null) {
                 $this->logger->info("No results parsed.");
-                return $this;
+                return;
             }
-            if (!is_null($bucketName)) {
-                $path .= $bucketName . '/';
-                $bucketName = $sapiPrefix ? 'in.c-' . $bucketName : $bucketName;
-            }
+
             if (!is_dir($path)) {
                 mkdir($path, null, true);
                 chown($path, fileowner($outdir));
@@ -139,14 +118,12 @@ class Processor
 
             $resFileName = $key . '.csv';
             $manifest = [];
-            if (!is_null($bucketName)) {
-                $manifest['destination'] = "{$bucketName}.{$key}";
-            }
-            $manifest['incremental'] = is_null($file->getIncremental()) ? $incremental : $file->getIncremental();
+
+            $manifest['incremental'] = $file->isIncrementalSet() ? $file->getIncremental() : $incremental;
             if (!empty($file->getPrimaryKey())) {
                 $manifest['primary_key'] = $file->getPrimaryKey(true);
             }
-            $this->logger->info("Writting reult file: " . $resFileName);
+            $this->logger->info("Writing result file: " . $resFileName);
             file_put_contents($path . $resFileName . '.manifest', json_encode($manifest));
             copy($file->getPathname(), $path . $resFileName);
         }
